refactoring and new api
