+++ .travis.yml
@@ -4,7 +4,6 @@ os:
   - linux
   - osx
 julia:
-  - 1.0
   - 1.3
   - nightly
 
@@ -14,8 +13,7 @@ script:
 
 jobs:
   allow_failures:
-    - julia: nightly
-    - julia: 1.0
+    - julia: nightly    
     - os: osx
   include:
     - stage: "Documentation"
