+++ Project.toml
@@ -6,6 +6,7 @@ version = "0.2.0"
 [deps]
 AxisArrays = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
 BinDeps = "9e28174c-4ba2-5203-b857-d8d62c4213ee"
+CFTime = "179af706-886a-5703-950a-314cd64e0468"
 ClimateBase = "35604d93-0fb8-4872-9436-495b01d137e2"
 ClimateTools = "4f4ee721-4970-5af2-8560-6c1d960e3231"
 Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
@@ -36,7 +37,6 @@ julia = "1.3"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
 
 [targets]
 test = ["Test"]
