+++ test/runtests.jl
@@ -17,64 +17,70 @@ C = load(filenc, "tas")
 regions = ["World", "Mollweide", "eck1", "eck2", "eck3", "eck4", "eck5", "eck6","Canada", "Quebec", "USA","NorthAmerica", "SouthAmerica", "Arctic", "Antarctic", "Africa", "Europe", "Asia", "auto"]
 
 for iregion in regions
-    status, figh = ClimatePlots.mapclimgrid(C, region=iregion);@test status == true; PyPlot.close()
+    status, figh = ClimatePlots.contourf(C, region=iregion);@test status == true; PyPlot.close()
+    status, figh = ClimatePlots.contour(C, region=iregion);@test status == true; PyPlot.close()
+    status, figh = ClimatePlots.pcolormesh(C, region=iregion);@test status == true; PyPlot.close()
 end
 
 # precip
 C = load(filenc, "pr", data_units="mm") + 2.0
 
 for iregion in regions
-    status, figh = ClimatePlots.mapclimgrid(C, region=iregion);@test status == true; PyPlot.close()
+    status, figh = ClimatePlots.contourf(C, region=iregion);@test status == true; PyPlot.close()
+    status, figh = ClimatePlots.contour(C, region=iregion);@test status == true; PyPlot.close()
+    status, figh = ClimatePlots.pcolormesh(C, region=iregion);@test status == true; PyPlot.close()
 end
-
-status, figh = ClimatePlots.mapclimgrid(C, start_date=(2000, 05, 16, 12), end_date=(2000, 05, 16, 12));@test status == true; PyPlot.close()
-
 
 # ua wind
 C = load(filenc, "ua")
-status, figh = ClimatePlots.mapclimgrid(C, level = 3);@test status == true; PyPlot.close() # given level
-status, figh = ClimatePlots.mapclimgrid(C);@test status == true; PyPlot.close() # feeding a 4D field
+status, figh = ClimatePlots.contourf(C, level = 3);@test status == true; PyPlot.close() # given level
+status, figh = ClimatePlots.contourf(C);@test status == true; PyPlot.close() # feeding a 4D field
 
-# Empty maps
-regions = ["World", "Mollweide", "eck1", "eck2", "eck3", "eck4", "eck5", "eck6","Canada", "Quebec", "USA","NorthAmerica", "SouthAmerica", "Arctic", "Antarctic", "Africa", "Europe", "Asia"]
-
-for iregion in regions    
-    status, figh = ClimatePlots.mapclimgrid(region=iregion);@test status == true; PyPlot.close()
-end
 
 # DUMMY MAPS
-lon = collect(-180.0:180.0)
-lat = collect(-90.0:90.0)
+lon = collect(-180.0:2.0:180.0)
+lat = collect(-90.0:2.0:90.0)
 longrid, latgrid = ndgrid(lon, lat)
 timeV = DateTime(2003,1,1):Day(1):DateTime(2003,01,31)
-data = randn(361, 181, 31)
+data = randn(181, 91, 31)
 dimension_dict = Dict(["lon" => "lon", "lat" => "lat"])
 varattribs = Dict(["standard_name" => "random noise"])
 axisdata = AxisArray(data, Axis{:lon}(lon), Axis{:lat}(lat), Axis{:time}(timeV))
 C = ClimateTools.ClimGrid(axisdata, variable = "psl", typeofvar="psl", longrid=longrid, latgrid=latgrid, dimension_dict=dimension_dict, varattribs=varattribs)
 
-status, figh = ClimatePlots.mapclimgrid(C);@test status == true; PyPlot.close()
-status, figh = ClimatePlots.mapclimgrid(C, center_cs=true, caxis=[200, 250]);@test status == true; PyPlot.close()
-status, figh = ClimatePlots.mapclimgrid(C, center_cs=true, surface=:pcolormesh);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.contourf(C);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.contour(C);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.pcolormesh(C);@test status == true; PyPlot.close()
+
+status, figh = ClimatePlots.contourf(C, center_cs=true, caxis=[200, 250]);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.contour(C, center_cs=true, caxis=[200, 250]);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.pcolormesh(C, center_cs=true, caxis=[200, 250]);@test status == true; PyPlot.close()
 
 
-lon = collect(-180.0:180.0)
-lat = collect(-90.0:90.0)
+status, figh = ClimatePlots.contourf(C, center_cs=true);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.contour(C, center_cs=true);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.pcolormesh(C, center_cs=true);@test status == true; PyPlot.close()
+
+
+lon = collect(-180.0:2.0:180.0)
+lat = collect(-90.0:2.0:90.0)
 longrid, latgrid = ndgrid(lon, lat)
 timeV = Dates.year(DateTime(2003, 01, 01)):Dates.year(1):Dates.year(DateTime(2004, 01, 01))
-data = randn(361, 181, 2)
+data = randn(181, 91, 2)
 dimension_dict = Dict(["lon" => "lon", "lat" => "lat"])
 varattribs = Dict(["standard_name" => "random noise"])
 axisdata = AxisArray(data, Axis{:lon}(lon), Axis{:lat}(lat), Axis{:time}(timeV))
 C = ClimateTools.ClimGrid(axisdata, variable = "psl", typeofvar="psl", longrid=longrid, latgrid=latgrid, dimension_dict=dimension_dict, varattribs=varattribs)
 
-status, figh = ClimatePlots.mapclimgrid(C);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.contourf(C);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.contour(C);@test status == true; PyPlot.close()
+status, figh = ClimatePlots.pcolormesh(C);@test status == true; PyPlot.close()
 
-lon = collect(-180.0:180.0)
-lat = collect(-90.0:90.0)
+lon = collect(-180.0:2.0:180.0)
+lat = collect(-90.0:2.0:90.0)
 longrid, latgrid = ndgrid(lon, lat)
 timeV = DateTime(2003, 01, 01):Day(1):DateTime(2004, 01, 01)
-data = randn(361, 181, 366)
+data = randn(181, 91, 366)
 dimension_dict = Dict(["lon" => "lon", "lat" => "lat"])
 varattribs = Dict(["standard_name" => "random noise"])
 axisdata = AxisArray(data, Axis{:lon}(lon), Axis{:lat}(lat), Axis{:time}(timeV))
